+++ arius-admin/arius-admin-common/src/main/java/com/didichuxing/datachannel/arius/admin/common/bean/po/stats/ClusterLogicStatsPO.java
@@ -31,7 +31,7 @@ public class ClusterLogicStatsPO {
     /**
      * docNu
      */
-    private double docNu;
+    private Long docNu;
 
     /**
      * 总的磁盘容量
